+++ src/views/overview.blade.php
@@ -5,10 +5,12 @@
 	@include('admin::partials.new')
 	<h1>{{ $heading }}</h1>
 
-	@if (Session::has('model.updated'))
-		<div id="js-alert" class="alert alert-success" data-alert="alert">
-			{{ Session::get('model.updated') }}
-		</div>
+	@if (Session::has('model'))
+		@foreach(Session::get('model') as $msg)
+			<div id="js-alert" class="alert alert-success" data-alert="alert">
+				{{ $msg }}
+			</div>
+		@endforeach
 	@endif
 
 	@include('admin::partials.sort-alert')
@@ -44,9 +46,11 @@
 						Edit
 					</a>
 
-					{{ Form::open(['class' => 'inline-form', 'action' => [ $controller . '@destroy', $row->getId() ], 'method' => 'DELETE']) }}
-						<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
-					{{ Form::close() }}
+					@if ($canDelete)
+						{{ Form::open(['class' => 'inline-form', 'action' => [ $controller . '@destroy', $row->getId() ], 'method' => 'DELETE']) }}
+							<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
+						{{ Form::close() }}
+					@endif
 				</div>
 			</li>
 		@endforeach