+++ src/views/partials/nested-item.blade.php
@@ -5,10 +5,11 @@
 							<i class="fa fa-pencil"></i>
 							Edit
 						</a>
-
-						{{ Form::open(['class' => 'inline-form', 'action' => array($controller . '@destroy', $row->getId()), 'method' => 'DELETE']) }}
-							<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
-						{{ Form::close() }}
+						@if ($canDelete)
+							{{ Form::open(['class' => 'inline-form', 'action' => array($controller . '@destroy', $row->getId()), 'method' => 'DELETE']) }}
+								<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
+							{{ Form::close() }}
+						@endif
 					</div>
 					<div class="nested-value">
 						<i class="fa fa-sort sorting-handle"></i>