+++ src/views/partials/save.blade.php
@@ -1,1 +1,8 @@
-<button class="btn btn-success pull-right space-left" type="submit"><i class="fa fa-save"></i>Save</button>
+<button class="btn btn-success pull-right space-left" type="submit" name="after_save" value="exit">
+	<i class="fa fa-save"></i>
+	Save and Exit
+</button>
+<button class="btn btn-success pull-right space-left" type="submit" name="after_save" value="continue">
+	<i class="fa fa-save"></i>
+	Save and Continue
+</button>